+++ packages/client/src/pages/admin/overview.queue-chart.vue
@@ -150,7 +150,7 @@ onMounted(() => {
 			layout: {
 				padding: {
 					left: 0,
-					right: 8,
+					right: 0,
 					top: 0,
 					bottom: 0,
 				},